(PC-34592)[API] refactor: split Algolia indexing concepts appart
The AlgoliaBackend used to index offers to Algolia had too many
concepts embedded within:
- the communication with Algolia
- the Redis queues and hashmaps management
- the object's serialization

This commit split those 3 area into different parts.
